requirements.txt
streamlit
langchain
langchain_core
langchain_community
langchain_text_splitters
langchain_groq
langchain_huggingface
pypdf
faiss_cpu
python_dotenv
ipykernel
pyttsx3
gtts